Asset upgrade
added new assets for bridge, water and forest. The camera is also added to a gimbal but have yet to add the code for touch input rotation (use two fingers to rotate). Also the walking tiles now use 4 directional Raycast to detect other walking tiles and adust the objects mesh so that they connect