level implementation (first phase)
Added a Y axis to the generator so now its spawns multiple levels stacked up on eachother. This project has base models for forest, water, bridge and hero character. (waaay to arab tho fam, change him up) A camera controller is also implemented to allow the player to rotate camera around the level and touch tiles near empty spot so that it moves to the empty spot (try to make this shit slide tho fam instead of clicking cuz it legit sucks lmfao)